pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aider-chat>=0.79.1",
    "chromadb>=0.6.3",
    "grandalf>=0.8",
    "langchain>=0.3.20",
    "langchain-chroma>=0.2.2",
    "langchain-core>=0.3.45",
    "langchain-google-firestore>=0.5.0",
    "langchain-google-genai>=2.1.0",
    "langchain-mistralai>=0.2.8",
    "langchain-openai>=0.3.8",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.0.1",
]
